requirements.txt
django>=5.0.6,<5.0.99
djangorestframework>=3.15.2,<3.15.99
psycopg>=3.1.19,<3.1.99
Pillow>=10.3.0,<10.3.99
whitenoise>=6.7.0,<6.7.99
drf-spectacular>=0.27.2,<0.27.99
factory-boy>=3.3.0,<3.3.99
django-extensions>=3.2.3,<3.2.99
flake8>=7.1.0,<7.1.99
isort>=5.13.2,<5.13.99
black==24.4.2
celery>=5.4.0,<5.4.99
coverage>=7.5.4,<7.5.99
django-otp>=1.5.0,<1.5.99
qrcode==7.4.2
django-redis>=5.4.0,<5.4.99
django-cors-headers>=4.4.0,<4.4.99
django-debug-toolbar>=4.4.2,<4.4.99
sentry-sdk>=2.7.0,<2.7.99
